Player basic behaviour
Player --> Added a collider && rigidbody to stop the player vs walls.
If the player has the max health, then it uses the big sword, otherwise the small one.
Walls --> Added a box colider and made a prefab
BasicEnemy --> No IA, it deals dmg if the player triggers the collider.
Apple --> Heals the player 1 hp.